vallejo/requirements.txt
curl_cffi
easyocr
Pillow
numpy
python-dateutil
